+++ src/routes/router.tsx
@@ -5,14 +5,13 @@ import LoginPage from '../pages/LoginPage'
 import MyPage from '../pages/MyPage'
 import AdminPage from '../pages/AdminPage'
 import UserListPage from '../pages/UserListPage'
-import _LoginForm from '../components/Test_Calendar/Login'
+
 
 function Router() {
   return (
     <BrowserRouter>
       <Routes>
         <Route path="/" element={<MainPage />} />
-        <Route path="/test" element={<_LoginForm />} />
         <Route path="/login" element={<LoginPage />} />
         <Route path="/signup" element={<SignupPage />} />
         <Route path="/mypage" element={<MyPage />} />
